requirements.txt
streamlit==1.33.0
pandas==2.2.2
requests==2.31.0
plotly==5.20.0
sentence-transformers==2.7.0
scikit-learn==1.4.2
python-dotenv==1.0.1
google-api-python-client==2.128.0
firebase-admin